dep-updates: Move import to pdf folder
